fixed outlook button by making is function the same as the apple button and changed button colours
this is because the link can only point to a specific outlook address, for example a university account, work account, personal account. there are endless variations of MS emails due to many different companies and schools. to work around this i made the outlook button behave the same as the apple calendar button for now